% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_converters.R
\name{renv_to_description}
\alias{renv_to_description}
\title{Converts a renv lockfile to a DESCRIPTION file}
\usage{
renv_to_description(lockfile)
}
\arguments{
\item{lockfile}{\itemize{
\item renv lockfile
}}
}
\value{
Writes or appends to a DESCRIPTION file in the current project
}
\description{
Converts a renv lockfile to a DESCRIPTION file
}
