% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_converters.R
\name{renv_to_install_r}
\alias{renv_to_install_r}
\title{Title}
\usage{
renv_to_install_r(lockfile, file_out = "install.R")
}
\arguments{
\item{lockfile}{\itemize{
\item must be a renv lockfile
}}

\item{file_out}{\itemize{
\item path and file name
}}
}
\value{
Side effect: writes a install.R compatible with repo2docker
}
\description{
Title
}
